Refatorar lógica de login: mover para auth.js e criar src/services/auth.js
